fix(sam-serverless-app): add type safety to runtimeMappings
This changes `runtimeMappings` from a `Map<string, RuntimeMapping>`
to a `{ [key in BlueprintRuntimes]: RuntimeMapping; }`.

This improves type-safety by having TypeScript enforce that the map
has a value for every key defined in BlueprintRuntimes, and thus we
can remove non-null assertions in some parts of the code.

Tested by building and publishing the template, then verifying that
the Create Project wizard renders as expected.