feat(workflows): Add bucket name param to cleanup workflow
Previously each synthesis of a blueprint using a cleanup workflow would
generate a different bucket name, causing conflicts when performing merges
during resynthesis. Add an optional parameter to the cleanup workflow to
set the bucket name. This way, a blueprint can store the random bucket name
as an option so that resynthesis can generate a workflow with the same bucket
name.