fix(blueprint): Detect binary files during three way merges (#410)
The three way merge is text based and won't produce meaningful results
when run on binary files. Use a heuristic similar to git of checking the
first 3000 bytes of the input files for null bytes. If there are any null
bytes, consider the file to be binary and fall back to returning the proposed
file.